Conflicts:
	examples/Makefile

+++ util/concurrent_arena.cc
@@ -1,4 +1,4 @@
-//  Copyright (c) 2013, Facebook, Inc.  All rights reserved.
+//  Copyright (c) 2011-present, Facebook, Inc.  All rights reserved.
 //  This source code is licensed under the BSD-style license found in the
 //  LICENSE file in the root directory of this source tree. An additional grant
 //  of patent rights can be found in the PATENTS file in the same directory.
@@ -36,12 +36,13 @@ ConcurrentArena::Shard* ConcurrentArena::Repick() {
   int cpuid = port::PhysicalCoreID();
   if (UNLIKELY(cpuid < 0)) {
     // cpu id unavailable, just pick randomly
-    cpuid = Random::GetTLSInstance()->Uniform(index_mask_ + 1);
+    cpuid =
+        Random::GetTLSInstance()->Uniform(static_cast<int>(index_mask_) + 1);
   }
 #if ROCKSDB_SUPPORT_THREAD_LOCAL
   // even if we are cpu 0, use a non-zero tls_cpuid so we can tell we
   // have repicked
-  tls_cpuid = cpuid | (index_mask_ + 1);
+  tls_cpuid = cpuid | (static_cast<int>(index_mask_) + 1);
 #endif
   return &shards_[cpuid & index_mask_];
 }
